requirements.txt
Flask
mysql-connector-python
python-dotenv
pandas
matplotlib
sqlalchemy
gunicorn
numpy